+++ src/main/resources/db/migration/V1__Create-Users-Table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE user(
-    id bigint NOT NULL PRIMARY KEY AUTO_INCREMENT,
+    id INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
     name varchar(255) DEFAULT NULL,
     last_name varchar(255) DEFAULT NULL,
     email varchar(255) DEFAULT NULL,
