+++ src/main/resources/db/migration/V3__Create-Expenses-Table.sql
@@ -5,7 +5,7 @@ CREATE TABLE expense(
     installment_count int DEFAULT NULL,
     purchase_date datetime(6) DEFAULT NULL,
     description varchar(255) DEFAULT NULL,
-    user_id bigint DEFAULT NULL,
+    user_id INT DEFAULT NULL,
     FOREIGN KEY (user_id) REFERENCES user(id)
     ON DELETE RESTRICT
     ON UPDATE RESTRICT
